+++ src/fr/adrienbrault/idea/symfony2plugin/config/PhpClassReference.java
@@ -1,0 +1,53 @@
+package fr.adrienbrault.idea.symfony2plugin.config;
+
+import com.intellij.codeInsight.lookup.LookupElement;
+import com.intellij.psi.*;
+import com.jetbrains.php.PhpIndex;
+import com.jetbrains.php.lang.psi.elements.PhpClass;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.*;
+
+/**
+ * @author Daniel Espendiller <daniel@espendiller.net>
+ */
+public class PhpClassReference extends PsiReferenceBase<PsiElement> implements PsiPolyVariantReference {
+
+    private String classFQN;
+
+    public PhpClassReference(@NotNull PsiElement element, String classFQN) {
+        super(element);
+        this.classFQN = classFQN;
+    }
+
+    @NotNull
+    @Override
+    public ResolveResult[] multiResolve(boolean incompleteCode) {
+
+        PhpIndex phpIndex = PhpIndex.getInstance(getElement().getProject());
+        Collection<PhpClass> phpClasses = phpIndex.getClassesByFQN(classFQN);
+
+        List<ResolveResult> results = new ArrayList<ResolveResult>();
+        for (PhpClass phpClass : phpClasses) {
+            results.add(new PsiElementResolveResult(phpClass));
+        }
+
+        return results.toArray(new ResolveResult[results.size()]);
+    }
+
+    @Nullable
+    @Override
+    public PsiElement resolve() {
+        ResolveResult[] resolveResults = multiResolve(false);
+        return resolveResults.length == 1 ? resolveResults[0].getElement() : null;
+    }
+
+    @NotNull
+    @Override
+    public Object[] getVariants() {
+        List<LookupElement> results = new ArrayList<LookupElement>();
+        // fill this
+        return results.toArray();
+    }
+}